+++ spec-integration/createPet.spec.js
@@ -1,5 +1,7 @@
+/* eslint-disable no-unused-expressions */
 const { expect } = require('chai');
 const sinon = require('sinon');
+const logger = require('@elastic.io/component-logger')();
 const action = require('../lib/actions/createPet');
 
 const cfg = {
@@ -14,9 +16,11 @@ const msg = {
   },
 };
 
-// we use sinon to mock functions such as `emit`, which come from sailor when the code is loaded onto the platform
+// we use sinon to mock functions such as `emit`,
+// which come from sailor when the code is loaded onto the platform
 const self = {
   emit: sinon.spy(),
+  logger,
 };
 
 describe('Get Pets By Status', () => {