+++ spec-integration/petstoreClient.spec.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-expressions */
 const { expect } = require('chai');
 const PetstoreClient = require('../lib/petstoreClient');