+++ verifyCredentials.js
@@ -1,7 +1,8 @@
 const MagentoClient = require('./lib/petstoreClient');
 
 /**
- * Executes the verification logic by sending a simple to the Petstore API using the provided apiKey.
+ * Executes the verification logic by sending a simple to the
+ * Petstore API using the provided apiKey.
  * If the request succeeds, we can assume that the apiKey is valid. Otherwise it is not valid.
  *
  * @param credentials object to retrieve apiKey from